Add GUI & filter with humorous description.
The GUI will be the point of interaction with the server-side checker
and data manager. As such it needs to be light and simple.

The humorous description needs to get the point across that this is a
joyful little project meant to poke fun at a semi-issue, while
encouraging participation.

The filter is quick and regex enabled. It does what we want it to, and
well. The display of subdomains as separate from ed.ac.uk also adds
clarity about the filter.