Add comprehensive UI and net communication
Comprehensive UI means that users should be able to more easily
understand the concept of the project and know what kinds of options are
available to them.

Net communication has linked the two halves of the brain. The core
functionality is now complete and properly operating from server to
client.

This also resolved some old bugs and generally cleaned the code up a tad
bit.