Add work towards curl utility. Don't implement.
Unfortunately, because of some strange problems on the side of the
servers for domains such as ed.ac.uk, http requests and curls never come
back. May be caused by an insistence on https?

This is irritating to say the least. Either way, the code has been
commented out such that it may be of use later once I've diagnosed the
problem.