Change client-side handling of lastCheck == 0
The client-side script, when making a request for a new domain, will
react to lastCheck == 0 on either ping or http and make requests
accordingly to the server. Thanks to the RemoteProperty class, these
trquests will be answered with initial checks run once the Domain object
is initialized on the server side, thereby preventing multiple requests
for one resource in a short amount of time.