Remove breaking bug from previous false searches
Previous false searches (those that didn't resolve to a domain name)
would still be added to the database and not removed after being
confirmed as non-existent. This creates a clash with the SQLITE
uniqueness property of the primary key when the same faulty domain tries
to be added later.

The solution is to remove the entry after it is confirmed as
non-existent.

This may not be the last fix.