Start progress towards faster UI.
Loads whole DB into UI in under second, but we've had to "cut through
bone" in a few places, metaphorically speaking.

The status display for more details will later be shown in a different
place than the list, and the list is now completely formatted as text
with minimal styling.

The server-side now implements its own sorting, so the client can ignore
it when adding large amounts of new rows. If more than 500 rows are
being added, the UI defaults to writing all changes to a string before
sending it into .innerHTML. This is faster for large payloads.

A fast UI is vital if we are to continue the project, seeing as there
are now so many domains, and the implementation of descriptions and
status history will only add more stress to the system.