Complete lightweight Ui with DomainInfo
Add DomainInfo singleton for controlling the output of additional info
about each domain. Some upgrades to the way info is shown and formatted.

Make sure UI is tight and responsive on mobile as well.

Teach other classes to link with and communicate with DomainInfo.

Otherwise, some small fixes here and there to recent changes.

This represents the last change for the lightweight UI. Yay!